refactor(converter): delete the superseded MarkdownConverter
internal/converter/markdown.go rendered a complete markdown report that the
binary never invoked. The shipped report comes from internal/converter/builder
via HybridGenerator; MarkdownConverter had no production caller, so a
contributor editing it saw no change in opndossier output.

Its Converter interface and the Adapter interface that embedded it go with it.
Adapter had no implementations at all -- its own comment said the real adapter
had moved elsewhere.

Tests are repointed, not dropped. The pfSense parser test asserting the report
title is device-type-aware now runs against the shipped generator, which emits
"# pfSense Configuration Summary", so the assertion holds and starts covering
the real path instead of a dead one. Benchmarks keep their names and measure
the same thing through the entry point production uses. The two bench helpers
that lived in the deleted file move to their remaining consumer.

One test is deleted rather than repointed: TestGetTheme_EdgeCases covered a
glamour-theme helper local to the deleted converter. The shipped path never
calls glamour.Render, and the production theme logic in internal/display has
its own coverage, so nothing is lost.

Golden fixtures are byte-identical.
